Implement `Notifier` and more
