+++ app/src/commonMain/kotlin/diy/lingerie/frp/ChangingCell.kt
@@ -1,9 +1,12 @@
 package diy.lingerie.frp
 
+import diy.lingerie.frp.Notifier.ListenerStrength
+import diy.lingerie.utils.iterable.forEachRemoving
+
 abstract class ChangingCell<V>(
     initialValue: V,
 ) : Cell<V>() {
-    private val changesEmitter = EventEmitter<Change<V>>()
+    private val changeListeners = mutableSetOf<Notifier.ListenerReference<Change<V>>>()
 
     private var mutableValue: V = initialValue
 
@@ -11,19 +14,43 @@ abstract class ChangingCell<V>(
         get() = mutableValue
 
     final override val changes: EventStream<Change<V>>
-        get() = changesEmitter
+        get() = ProxyEventStream(source = this)
 
     protected fun update(newValue: V) {
         val oldValue = mutableValue
 
+        val change = Change(
+            oldValue = oldValue,
+            newValue = newValue,
+        )
+
         mutableValue = newValue
 
-        changesEmitter.emit(
-            Change(
-                oldValue = oldValue,
-                newValue = newValue,
-            ),
-        )
+        changeListeners.forEachRemoving { listenerReference ->
+            listenerReference.handle(change)
+        }
+    }
+
+    override fun subscribe(
+        listener: Listener<Change<V>>,
+        strength: ListenerStrength,
+    ): Subscription {
+        val listenerReference = strength.refer(listener)
+
+        val wasAdded = changeListeners.add(listenerReference)
+
+        if (!wasAdded) {
+            throw AssertionError("Value listener reference was already present (???)")
+        }
+
+        return object : Subscription {
+            override fun cancel() {
+                val wasRemoved = changeListeners.remove(listenerReference)
+
+                if (!wasRemoved) {
+                    throw IllegalStateException("New value listener was not found")
+                }
+            }
+        }
     }
 }
-
