+++ app/src/commonMain/kotlin/diy/lingerie/frp/ConstCell.kt
@@ -6,4 +6,9 @@ internal class ConstCell<V>(
     override val currentValue: V = constValue
 
     override val changes: EventStream<Nothing> = EventStream.Never
+
+    override fun subscribe(
+        listener: Listener<Change<V>>,
+        strength: Notifier.ListenerStrength
+    ): Subscription = Subscription.Noop
 }
