+++ app/src/commonMain/kotlin/diy/lingerie/frp/NeverEventStream.kt
@@ -3,6 +3,6 @@ package diy.lingerie.frp
 internal object NeverEventStream : EventStream<Nothing>() {
     override fun subscribe(
         listener: Listener<Nothing>,
-        strength: NotifyingStream.ListenerStrength
+        strength: Notifier.ListenerStrength
     ): Subscription = Subscription.Noop
 }
