+++ app/src/commonTest/kotlin/diy/lingerie/frp/EventStreamVerifier.kt
@@ -1,5 +1,7 @@
 package diy.lingerie.frp
 
+import diy.lingerie.frp.subscribeSemiBound
+
 class EventStreamVerifier<E>(
     eventStream: EventStream<E>,
 ) {
